introduce data/cutoff-dates.yaml, and move some functions to utils.py
