+++ py/fiberassign/utils.py
@@ -11,6 +11,9 @@ from __future__ import absolute_import, division, print_function
 
 import os
 import sys
+from datetime import datetime
+import yaml
+from pkg_resources import resource_filename
 
 from ._internal import (Logger, Timer, GlobalTimers, Circle, Segments, Shape,
                         Environment)
@@ -60,3 +63,40 @@ def option_list(opts):
                 else:
                     optlist.append("{}".format(val))
     return optlist
+
+
+def assert_isoformat_utc(time_str):
+    """
+    Asserts if a date formats as "YYYY-MM-DDThh:mm:ss+00:00".
+
+    Args:
+        time_str: string with a date
+    Returns:
+        boolean asserting if time_str formats as "YYYY-MM-DDThh:mm:ss+00:00"
+    """
+    try:
+        test_time = datetime.strptime(time_str, "%Y-%m-%dT%H:%M:%S%z")
+    except ValueError:
+        return False
+    # AR/SB it parses as an ISO string, now just check UTC timezone +00:00 and not +0000
+    return time_str.endswith("+00:00")
+
+
+def get_date_cutoff(datetype, cutoff_case):
+    """
+    Returns the cutoff date for a particular case, from reading the data/cutoff-dates.yaml file.
+
+    Args:
+        datetype: one of the key of data/cutoff-dates.yaml (e.g., "rundate", "mtltime", "pmtime_utc_str") (string)
+        cutoff_case: one of the config[datetype] keys in data/cutoff-dates.yaml, (e.g. "std_wd" for datetype="rundate") (string)
+
+    Returns:
+        date_cutoff: cutoff date for datetype and cutoff_case, in the "YYYY-MM-DDThh:mm:ss+00:00" formatting (string)
+
+    """
+    fn = resource_filename("fiberassign", "data/cutoff-dates.yaml")
+    with open(fn, "r") as f:
+        config = yaml.safe_load(f)
+    f.close()
+    date_cutoff = config[datetype][cutoff_case]
+    return date_cutoff
